dom/html/test/file_fullscreen-plugins.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=545812

Test plugins with DOM full-screen API:
* Request for full-screen is denied when windowed plugin in current doc is present.
* Request for full-screen is denied when windowed plugin in subdocument is present.
* Request for full-screen is not denied when the only plugin present is windowless.
* Adding an existing (out-of-doc) windowed plugin to a full-screen document causes document to exit full-screen.
* Create windowed plugin and adding it to full-screen document caused exit from full-screen.
* On non-Windows, plugins can only be windowless, so the presence of plugins
  should have no effect on request for full-screen.

-->
<head>
  <title>Test for Bug 545812</title>
  <script type="application/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="plugin-utils.js"></script>
  <script type="application/javascript">
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);
  </script>
  <script type="application/javascript" src="file_fullscreen-utils.js"></script>
  <style>
  body:fullscreen, div:fullscreen {
    background-color: red;
  }
  </style>
</head>
<body>


<!-- Windowed plugin, focusing should revoke full-screen. -->
<embed id="windowed-plugin" type="application/x-test" style="width:200px;height:100px;" wmode="window"></embed>

<!-- Windowless plugin, focusing should not revoke full-screen. -->
<embed id="windowless-plugin" type="application/x-test" style="width:200px;height:100px;"></embed>


<!-- iframe contents:

<html><body><embed id='windowed-plugin' type='application/x-test' style='width:200px;height:100px;' wmode='window'></embed></body></html>

-->

<iframe id="subdoc-plugin" src="data:text/html;charset=utf-8,<html><body><embed id%3D'windowed-plugin' type%3D'application%2Fx-test' style%3D'width%3A200px%3Bheight%3A100px%3B' wmode%3D'window'><%2Fembed><%2Fbody><%2Fhtml>%0D%0A"></iframe>

<script type="application/javascript">

/** Test for Bug 545812 **/

function ok(condition, msg) {
  opener.ok(condition, "[plugins] " + msg);
}

function is(a, b, msg) {
  opener.is(a, b, "[plugins] " + msg);
}

function e(id) {
  return document.getElementById(id);
}

function removeElement(e) {
  e.remove();
}

const supportsWindowedMode = navigator.appVersion.indexOf("Windows") != -1;

var windowedPlugin = null;

function begin() {
  // Delay test startup long enough for the windowed plugin in the subframe to
  // start up and create its window.
  opener.SimpleTest.executeSoon(function() {
    opener.SimpleTest.executeSoon(function() {
      startTest();
    })
  });
}

function startTest() {
  ok(!document.fullscreenElement, "Should not be in full-screen mode initially");
  document.body.requestFullscreen();

  // Focus the windowed plugin. On non-Windows we should still enter
  // full-screen mode, on Windows the pending request for full-screen should
  // be denied.
  e("windowed-plugin").focus();

  if (!supportsWindowedMode) {
    // If all plugins are effectively windowless, request for full-screen should be granted.
    // Continue test in the "fullscreenchange" handler.
    addFullscreenChangeContinuation("enter", windowlessFullScreenChange1);
  } else {
    // On Windows, request should be denied, carry on the test after receiving error event.
    addFullscreenErrorContinuation(windowsTest);
  }
}

function windowsTest() {
  ok(!document.fullscreenElement, "Request for full-screen with focused windowed plugin should be denied.");

  // Focus a regular html element, and re-request full-screen, request should be granted.
  e("windowless-plugin").focus();
  addFullscreenChangeContinuation("enter", windowsTest2);
  document.body.requestFullscreen();
}

function windowsTest2() {
  ok(document.fullscreenElement, "Request for full-screen with non-plugin focused should be granted.");
  // Focus a windowed plugin, full-screen should be revoked.
  addFullscreenChangeContinuation("exit", windowsTest3);
  e("windowed-plugin").focus();
}

function windowsTest3() {
  ok(!document.fullscreenElement, "Full-screen should have been revoked when windowed-plugin was focused.");
  // Remove windowed plugins before closing the window
  // to work around bug 1237853.
  removeElement(e("windowed-plugin"));
  removeElement(e("subdoc-plugin").contentDocument.getElementById("windowed-plugin"));
  opener.nextTest();
}

var fullScreenChangeCount = 0;

function createWindowedPlugin() {
  var p = document.createElement("embed");
  p.setAttribute("type", "application/x-test");
  p.setAttribute("wmode", "window");
  return p;
}

function windowlessFullScreenChange1(event) {
  ok(document.fullscreenElement, "Requests for full-screen with focused windowed plugins should be granted on non-Windows");
  
  // Create a new windowed plugin, and add that to the document. Should *not* exit full-screen mode on MacOS/Linux.
  windowedPlugin = createWindowedPlugin();
  document.body.appendChild(windowedPlugin);
  
  // Focus windowed plugin. Should not exit full-screen mode on MacOS/Linux.
  windowedPlugin.focus();
  
  setTimeout(
    function() {
      ok(document.fullscreenElement, "Adding & focusing a windowed plugin to document should not cause full-screen to exit on MacOS/Linux.");
      addFullscreenChangeContinuation("exit", windowlessFullScreenChange2);
      document.exitFullscreen();
    }, 0);
}

function windowlessFullScreenChange2(event) {
  ok(!document.fullscreenElement, "Should have left full-screen mode after calling document.exitFullscreen().");
  opener.nextTest();
}

</script>
</pre>
</body>
</html>
